National Stock Price can plot data of stocks of a particular company allowing study of variation over the year.
National Stock Price, extract data from Quandl.com and pickles it in our
computer hard disk. Then we can plot a well maintained graph out of it.
Through this graph we can study the stock variations of a particular
company over the years.

This project is made in python, using popular libraries pandas and
matplotlib. To extract data from Quandl their API is used. Pickling
helps the user to store extracted data in the hard disk and use it later
for plotting the graph, this will prevent from downloading data every
time in a process which takes time.

You can later update the downloaded data and then plot it again with the
updated values.

To get the name of the company whose data you wish to plot, one needs to
go to quandl.com and search for "Indian NSE", this will provide name of
the company along with their "Quandl Code". Just enter the Quandl Code
in the project and you will be able to store and plot the data.

As per the graph it contains three partitions, one which uses "Line
Graph" to show the "Opening value" of the stock, second is a "Candlstick
graph" that calculates the total price from opening, closing, high, low,
last, total trade and turnover. Third is two "Line Graph", of "Closing
value" which shows the profit or lose of the stock at the closing of
stock.